api/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
beautifulsoup4
playwright
azure-cognitiveservices-vision-computervision
google-cloud-vision
azure-keyvault-secrets
azure-identity